+++ src/unpoly-migrate/motion.js
@@ -1,0 +1,6 @@
+let modernFinish = up.motion.finish
+
+up.motion.finish = function(...args) {
+  modernFinish(...args)
+  return up.migrate.formerlyAsync('up.motion.finish()')
+}