+++ src/unpoly/motion.js
@@ -365,7 +365,7 @@ up.motion = (function() {
 
   > [IMPORTANT]
   > The `up:motion:finish` event does **not** signal the end of an animation.
-  > For this see [awaiting postprocessing](/render-lifecycle#awaiting-postprocessing).
+  > For this see [awaiting postprocessing](/render-lifecycle#postprocessing).
 
   @event up:motion:finish
   @param {Element} event.target